chore: handle_conflict returns DecisionLevel

+++ src/solver/search.rs
@@ -252,7 +252,7 @@ fn search(
             std::cmp::Ordering::Greater => {
                 ruduction_pressure += 1;
                 processing_pressure += 1;
-                cdb.lbd.update(lbd);
+                cdb.lbd.update(lbd as u16);
             }
         }
         if ruduction_pressure >= processing_interval {
